Refactor lib.rs into modular structure
Break down monolithic lib.rs (2,866 lines) into 13 well-organized modules to improve maintainability and code organization. The refactoring creates 6 top-level modules (error, control, types, sdk, filter_wheel, backend) and 7 camera sub-modules organized by functional responsibility (lifecycle, configuration, readout_modes, info, imaging, parameters).

All tests passing (98/98). The public API remains 100% backward compatible with no breaking changes. Updated design documentation to reflect the new module structure.
